feat(front): add preview component
