+++ src/types/componentTypes.ts
@@ -1,0 +1,5 @@
+import PreviewComponent from '@/components/media/upload/PreviewComponent.vue'
+
+export interface PreviewComponentRef extends InstanceType<typeof PreviewComponent> {
+  emptyMedia: () => void
+}
